ML/requirements.txt
# Core ML/CV
# face-recognition==1.3.0
opencv-python==4.8.1.78
dlib-bin
numpy
Pillow

# Database
mysql-connector-python
dotenvy

# Utilities
python-dateutil

# API
fastapi
uvicorn[standard]==0.25.0
python-multipart==0.0.6
